feat(backend): W.O., court assignment, standings tournamentPoints, bracket results and FAP seed templates (33–48 pairs)

+++ backend/src/main/java/com/padeladmin/padeladmin/dto/pair/PairResponseDto.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class PairResponseDto {
     private Long id;
     private Long tournamentId;
-    private Integer totalPoints;
+    private Double totalPoints;
     private List<PairPlayerResponseDto> players;
     private List<PairScheduleConstraintResponseDto> constraints;
 }
